// src/__generated__/seeCategories.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: seeCategories
// ====================================================

export interface seeCategories_seeCategories {
  __typename: "Category";
  id: number;
  name: string;
}

export interface seeCategories {
  seeCategories: (seeCategories_seeCategories | null)[] | null;
}